Moving all detectors into a 'detectors' module
All detectors will be in 'org.firstinspires.ftc.teamcode.detectors' to clean up and organize
